Fix SQL issue and allow saving nicknames that are not currently on your player
