+++ src/main/java/simplexity/simplenicks/commands/subcommands/basic/DeleteSubCommand.java
@@ -1,21 +1,17 @@
-package simplexity.simplenicks.commands.subcommands;
+package simplexity.simplenicks.commands.subcommands.basic;
 
 import com.mojang.brigadier.Command;
 import com.mojang.brigadier.builder.LiteralArgumentBuilder;
 import com.mojang.brigadier.context.CommandContext;
 import com.mojang.brigadier.exceptions.CommandSyntaxException;
-import com.mojang.brigadier.exceptions.SimpleCommandExceptionType;
 import io.papermc.paper.command.brigadier.CommandSourceStack;
 import io.papermc.paper.command.brigadier.Commands;
-import io.papermc.paper.command.brigadier.MessageComponentSerializer;
-import net.kyori.adventure.text.minimessage.MiniMessage;
-import net.kyori.adventure.text.minimessage.tag.resolver.Placeholder;
 import org.bukkit.OfflinePlayer;
 import org.bukkit.entity.Player;
 import org.jetbrains.annotations.NotNull;
-import simplexity.simplenicks.SimpleNicks;
 import simplexity.simplenicks.commands.NicknameProcessor;
 import simplexity.simplenicks.commands.arguments.NicknameArgument;
+import simplexity.simplenicks.commands.subcommands.Exceptions;
 import simplexity.simplenicks.config.Message;
 import simplexity.simplenicks.saving.Nickname;
 import simplexity.simplenicks.util.Constants;
@@ -23,16 +19,6 @@ import simplexity.simplenicks.util.Constants;
 @SuppressWarnings("UnstableApiUsage")
 public class DeleteSubCommand implements SubCommand {
 
-    private final MiniMessage miniMessage = SimpleNicks.getMiniMessage();
-
-    private final SimpleCommandExceptionType ERROR_CANNOT_DELETE = new SimpleCommandExceptionType(
-            MessageComponentSerializer.message().serialize(
-                    miniMessage.deserialize(
-                            Message.ERROR_DELETE_FAILURE.getMessage(),
-                            Placeholder.parsed("prefix", Message.PLUGIN_PREFIX.getMessage())
-                    )
-            )
-    );
 
     @Override
     public void subcommandTo(@NotNull LiteralArgumentBuilder<CommandSourceStack> root) {
@@ -56,9 +42,8 @@ public class DeleteSubCommand implements SubCommand {
             if (player instanceof Player onlinePlayer) sendFeedback(onlinePlayer, Message.DELETE_NICK, nickname);
             return Command.SINGLE_SUCCESS;
         }
-        throw ERROR_CANNOT_DELETE.create();
+        throw Exceptions.ERROR_CANNOT_DELETE.create();
     }
-
 
 
     @Override
